LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/forward-to-pruned-entry-expected.txt


FAIL If forward pruning clobbers the target of a traverse, abort assert_equals: expected 1 but got 2
